modules/trufflehog_module/requirements.txt
redis>=4.0.0
aiofiles>=0.8.0
python-magic>=0.4.27 